Rewrite tech docs intro to refer to GOV.UK Docker
This follows the approach we take in other repos [1]. Note that in
development, Smart Answers can still be previewed locally - '--live'
just helps to avoid the need for other dependencies, like Static.

I think the description that Smart Answers contains and presents its
own content is covered in the intro paragraph ("self contained").

[1]: https://github.com/alphagov/collections/pull/2375