Rewrite MarriageAbroad flow test
This ports the Marriage Abroad Flow tests to the style proposed in [1]
and was done as part of the initial experimentation work [2]. It then
deletes the previous integration and unit tests.

While creating this I discovered that, by the old tests not rendering
the views, that there was a missing stub for Sweden organisation which
I've added in.

I did do some optimisations to speed up these new tests. I found that by
stubbing only the worldwide API for the country under test, that there
was a significant increase in performance.

Before:

```
Finished in 210.360104s, 16.6001 runs/s, 49.8193 assertions/s.
3492 runs, 10480 assertions, 0 failures, 0 errors, 0 skips
```

After:

```
Finished in 39.723775s, 88.0077 runs/s, 176.4938 assertions/s.
3496 runs, 7011 assertions, 0 failures, 0 errors, 0 skips
```

There is still a bunch of low hanging fruit that could increase the
performance of these tests. There are (somewhat depressingly) 1,275 near
identical view files with the copy: 'It’s not possible to get legal
recognition for your same-sex relationship in $country_name', these
could be converted to a generic partial to save perhaps a third of the
tests ran here from needing to be run. I imagine there are various
optimisations available for countries with near identical copy.

[1]: https://github.com/alphagov/smart-answers/pull/5459
[2]: https://github.com/alphagov/smart-answers/pull/5440