Rewrite Benefits Cap Calculator flow test
This converts the Benefits Cap Calculator flow tests to use the node based
testing approach introduced in [ADR-3][1].

This flow is somewhat unique in that it has a dynamic question system - which
needed considering in how to test it.

In putting this together there were some decisions to make about what
outcome logic to test as there is a setting up of values in the
calculator that is then output on the outcome. I went quite modest here
by just testing the benefit calculation logic. I could be persuaded that
it would be good to test the figures are outcome, but these are a little
annoying to test as the same thing is output on four different outcomes.

[1]: https://github.com/alphagov/smart-answers/blob/main/docs/arch/003-testing-flows.md