Parallelize tests based on number of processors
Rather than hard coding the number of workers that this should run on
this option uses the number of cores on the executing machine to decide
the distribution. At the time of me setting the original option I didn't
know that this option was available.

For most GOV.UK users this shouldn't make much difference as GDS issued
Macbook Pro's have 6 physical coress. But this should be more friendly to
any configurations that have a differing number of cores.