Demonstrate testing approach for a shared Smart Answer
This applies the new flow testing approach to the
am-i-getting-minimum-wage and minimum-wage-calculator-employers Smart
Answers. These flows are somewhat special, in the world of Smart
Answers, as they share most of their questions.

I've wrote this as a study as to how we'd approach testing the flow of a
Smart Answer that shares nodes while trying to avoid repeating tests.

The approach I've used is to include the test through helper methods.
This has the advantage of very low repetition between the files, but has
the disadvantage that the tests aren't re-runnable based on the line
they fail on. I can't think of a good way to solve this and think it
might just be a pain point that needs to be lived with for shared
minitests.

Both of these flows had very similar, but ultimately different outcome
pages so I've put together individual tests for these in each flow.