Allow Smart Answer nodes to have a defined template name
Smart Answer nodes typically have a template file that matches the name
of the question. This commit provides a node with a `template_name` DSL
method that allows the setting and retrieval of a node's template name
(in a similar style to Flow methods [1]).

This has been done to support a quirky situation for the Child Benefit
Calculator's Flow. This Flow has a dynamic question process where there
are a variable number of questions based on how many children you
answer. To resolve this I intend to create multiple questions that use
the same template and thus need a way to specify a common file used.

I've amended some of the tests for view template path to be clearer that
they expect a path to try provide more distinction as to the purpose of
that compared to the purpose of this.

[1]: https://github.com/alphagov/smart-answers/blob/59753722534a1f7d4d361bcf57bc1fa69f8a484f/lib/smart_answer/flow.rb#L27-L40