Set responses in a hash for testing flows
This changes the approach to setting responses for test flows to use a
hash rather than an array. This has been done as it helps isolate an
area where tests were passing despite users being unable to complete a
flow.

The scenario this occurred in was one where a Smart Answer flow was
sending users to the same question multiple times as a way to get
different answers in different contexts. This was not a supported
feature of Smart Answers and is not compatible with the migration path
to query string or session flows. It recently broke with the change in
PR #5357 [1].

The side-effects of setting @response to a hash is that adding a
invalid response and then a subsequent response will mean that you're
just re-answering the first question. Because of this the test in
SimplifiedExpensesCheckerTest needed to be amended.

[1]: https://github.com/alphagov/smart-answers/pull/5337