Remove assertion-less test
This is a rather strange test because it doesn't have any assertions in
it. I found that my changes in the previous commit broke the test and
when I looked at it I couldn't really make sense of what it was actually
testing (if anything).

As this is only a test for test helpers it doesn't seem much of a
casualty to delete.