Use child_number instead of child_index
This changes the way data is stored in the calculator for
ChildBenefitTaxCalculator so that it uses a variable named child_number
rather than child_index and starts at the number 1.

This has been done for greater consistency with the numbers used to
define the questions so that they share the same variable name and the
numbers correspond. Previously questions for child 4 were stored in an
array in index of 3 for instance.