Replace dynamic questions of ChildBenefitTaxCalculator
The Child Benefit Tax Calculator has a quirky feature where a user can
be asked the questions about child benefit start and end dates multiple
times, depending on the number of children a user has. This feature
relied on there not being a mapping between a question and an answer, as
the same question would have multiple different answers. This
unfortunately meant that it wasn't compatible with the recently
introduced response store approach to Smart Answers [1] and was broken
by work to consolidate the two approaches to Smart Answer response
storage [2].

The options to resolve this were rather limited, as there is no
functionality to dynamically define questions for a Smart Answer (nor
does this seem simple as the object lifecycle of a Flow would need to be
reconsidered). The approach I took to resolve this was to utilise the
maximum number of children to define a node per child. This allowed
there to no longer be dynamic nodes. This is perhaps a bit of a cop-out
as it may be worth investigating whether dynamic nodes should be a
feature, however as this is broken in production it was best to be
pragmatic.

[1]: https://github.com/alphagov/smart-answers/blob/59753722534a1f7d4d361bcf57bc1fa69f8a484f/docs/arch/002-store-responses-in-query-string.md
[2]: https://github.com/alphagov/smart-answers/pull/5337