Add dynamic options to Radio and Checkbox questions
This adds the ability to specify a block to resolve option keys are
request time. The block has the same context as other blocks in flow
definitions (e.g. on_response, next_node). The block should return an
array of string which are keys of the options to display. Specifying
static options is still available, however an option block takes
precedence if both are specified.