Remove unconventional SmartAnswer namespace from flows
Flow classes have been defined within a SmartAnswer namespace despite
them not residing within the appropriate directories for them to be
autoloaded. This has led to some quirky code to work around this (such
as manually requiring files and relying on `load` as a way to fake class
reloading).

As a precursor to the decision in ADR-3 [1] to move these files to
app/flows this PR removes this SmartAnswer namespace so these files no
longer have a shared namespace, as you'd expect of code in app/*
directories. Removing this namespace also has the benefit of making this
application business logic distinct from the SmartAnswer library logic
that powers the flows.

Arguably it'd also make sense to move Calculators out of the SmartAnswer
namespace also, but that's not a yak to shave today.

I've also removed the Shared namespace that was set on the files in
lib/smart_answer_flows/shared. With the removal of SmartAnswer these
would have been left in a global Shared namespace, which seemed quite
surprising. There doesn't seem to be any need for the Shared namespace
so I'm interpreting that as a directory purely for file organisation and
thus doesn't need code representation.

[1]: https://github.com/alphagov/smart-answers/blob/main/docs/arch/003-testing-flows.md