Removing the id parameter constraint
This constraint was originally added to prevent a 400 (Bad request) error from being returned to the user by masking it as a 404. However, this isn't appropriate behaviour as the error type is correct and returning 404 could lead to confusion. This behaviour is also redundant as the current GOV.UK architecture prevents bad requests from reaching Smart Answers in the first place. Router would return a 404 as there would be no matching route.