Add FlowTestHelper
This is a helper file for Flow tests which enables the Flow test syntax
proposed in ADR-3 [1].

This module contains helper methods to initialise the testing of a flow,
assertion helper methods to minimise repetition in Flow tests and a
class to represent a test flow.

The class is a refactoring of the various helper methods in
FlowIntegrationTestHelper [2] into a class form (it was previously a
variety of state setting methods mixed into the class). I felt a
dedicated class made this easier to understand and to get a feel for any
sprawl growing within it.

A method added that is worth noting is `ensure_valid_and_correct_node`.
This is used as a mechanism to raise an error if a Flow is invalid or is
not testing the node that would be expected. This has been added as a
result of my early work in setting up reworked Flow test and discovering
how easy it is to have an error in set-up and this then to affect any
assertions (can cause hard to debug issues or, worse, false positives).
This method has been added to run before common helper assertions as a
pre-condition before trying to assert.

[1]: https://github.com/alphagov/smart-answers/blob/main/docs/arch/003-testing-flows.md
[2]: https://github.com/alphagov/smart-answers/blob/210a097b941e81acec2aa6ecc1fc5007f115c15c/test/integration/smart_answer_flows/flow_integration_test_helper.rb