Configure Zeitwerk autoloader to load Flow classes
This adds some configuration to allow the Zeitwerk autoloader to load
the Flow classes automatically, doing so will remove the need to
manually require or `load` [1] these files.

The first quirk dealt with is that these classes have an unconventional
filename that doesn't match the class name (it uses a slug).

The second is that the files are in a lib/smart_answer_flows directory
which would normally be expected to be within a SmartAnswerFlows
namespace.

The third quirk is that there is also a shared directory within
lib/smart_answer_flows but code within there doesn't use a specific
namespace.

In the work to implement ADR-3 [2] I'm expecting to resolve both of
these first two quirks by moving these classes to app/flows and removing
the dash separated names. The third one is likely to remain, however
this doesn't seem to big a concern as it's a similar behaviour to
`concerns`.

[1]: https://github.com/alphagov/smart-answers/blob/e36f68bedac1e35007f6e07dbbb7b2da95494be7/lib/smart_answer/flow_registry.rb#L56-L60