Replace FlowRegistry needing to require/load flows
Now that Flow classes are autoloadable the manual load/require in
FlowRegistry behaviour is not needed. This behaviour was somewhat
undesirable in dev as it relied on the overridding method names as a
means to reload Flow classes when there was a dev change. This system
wasn't smart enough to allow the auto-reload of any shared data in a
SmartAnswer Flow.

By removing this we still need the fixture flows to be available as
these are not in the autoload path. To resolve this, these are manually
required when the method to set-up fixture flows is called.