Rename date fixture flow
This takes the rather scarily named
SmartAnswersControllerSampleWithDateQuestionFlow and changes it to
SalarySampleFlow.

This has been done as part of some clean-up work to increase consistency
in the fixture flows and remove the very long names given to some that
all start with smart-answers-controller.

It also makes the test a little simpler as the require and content store
stub are not necessary.