Replace FlowWithLinksFlow with RadioSampleFlow
This didn't need a special flow so I was able to remove the rather
FlowWithLinksFlow. The flow it replaces was rather odd as it didn't
actually have a flow class and was just a rogue template.

I'm a bit confused by what is going on in the BrokenLinkReport tests,
there's surely a simpler way to unit test this class.