Update template directory test for new flow path
This removes the test for a flow with no name, since that is no longer a
possible situation (as the path is derived from the class name). It then
adapts the test for the flow path to utilise the class name.

I decided to use a stub as the default name of SmartAnswer::Flow seemed
a little confusing.