Add new question and outcome to marriage abroad flow
Japan Vietnam and Panama offer consular opposite sex civil partnerships.

To provide this information in the marriage abroad smart answer we have:
- Added a new question to ask whether couples want to get married or
register a civil partnership
- Added a new outcome for each of these countries

The outcomes have been placed in the directory for each of these
countries' outcomes. This is to avoid duplicating this outcome multiple
times and to make it easier to keep it consistent across all the
outcomes for each country.

Trello card:
https://trello.com/c/b6vvMECA/2600-2-add-new-question-and-outcome-to-japan-panama-and-vietnam-marriage-abroad-about-opposite-sex-civil-partnership