Use snake case naming approach for Flow classes
This replaces the use of slugs as a means to organise flow files and
instead uses the conventional Ruby snake case.

This has been done to make this codebase more conventional, which will
allow the removal of custom autoloader logic.

This was done with the following command:

for f in *.rb; do
  mv "$f" "$(echo "$f" | rev | cut -c 4- | rev | tr '-' '_')_flow.rb"
done