Rename Smart Answer flow directories
This applies the renaming of the Flow classes to the directories within
lib/smart_answer_flows. These are now named in the conventional Ruby
snake_case style and reflect the name of the classes.

These are slightly strange directories as they're mostly used for
storing erb files so probably shouldn't be in lib (but may not be
applicable for views). However they're also sometimes used for nested
classes where Ruby directory structure is needed for autoloading.