Convert Find Coronavirus Support flow tests
Converts the Find Coronavirus Support flow tests to use the
node based testing approach introduced in [ADR-3][1].

Some of the links are also being explicitly tested in the outcome as
much of the conditional guidance, especially for England and Wales
has the same link text.

Also removes the existing RSpec feature tests  and integration tests
as they are no longer needed.

[1]: https://github.com/alphagov/smart-answers/blob/main/docs/arch/003-testing-flows.md