Set additional parameters on state
This allow additional parameters and their values to be set on the state
object accessible within the flow. This allows additional parameters to
be used in calculator objects. This is needed the next steps for you
business flow - we need to access the `ct` parameter to hide or show a
result.