Add flow logic for `what_type_of_work?`
Adds logic that uses the question and outcomes added in the previous
commits to enable users to select what profession they work in, which
will point them to the most relevant visa information outcome.

Users will only be directed to these outcomes if they plan to work in
the UK for longer than 6 months.