Replace minimum wage flow tests
This applies the flow test approach introduced in ADR-3 [1] to the two
minimum wage flows. It removes the previous unit and integration flow
tests.

The minimum wage flow is rather special in that it is essentially the
same flow for two different Smart Answers with different templates for
each of them at the outcome stage. This makes it a bit of a pain to test
as it is superfluous to write the same tests out twice but we should
still execute the flow through all of the distinct files (mostly
templates).

The approach I've taken is to include the test through helper methods.
This has the advantage of very low repetition between the files, but has
the disadvantage that the tests aren't re-runnable based on the line
they fail on. I can't think of a good way to solve this and think it
might just be a pain point that needs to be lived with for shared
minitests.

Both of these flows had very similar, but ultimately different outcome
pages so I've put together individual tests for these in each flow.

[1]: https://github.com/alphagov/smart-answers/blob/main/docs/arch/003-testing-flows.md