Add ability to define setup block for flow definitions
This extends flow definitition to be able to run a block to setup the
starting state for a flow. Previously many flows setup the calculator
object in the first on_response block. This is was confusing as this
logic was not related to handling the response. Having an explicit setup
block allows logic to be executed even before the first node is
rendered. This is needed to for dynamic options, as existing flows
depend on intializing a calculator object to resolve their options. This
needs to happen before the first node is rendered and a response is
given.