Add setup hook for initializing nodes
This adds a setup method to run steps to correctly initialize a node.
This is run before just before the node is visited in the state
resolution process. Setup is used by radio and checkbox questions to
resolve dynamic option keys, moving this previously from the presenter
classes. This is because the presenter classes aren't created unless that
particular node is being rendered. This means option keys aren't
resolved when resolving state.