Use marriage abroad config in tests
In the review for this flow rewrite it was pointed out that one of the
constants was out of sync with the YAML config file (Congo was missing).
Rather than just fix the constant it seemed better to use the YAML file
as the means to power the tests, since it essentially just duplicates
the data.

To do this I've added some helper functions. I've also moved the helper
functions to the top of the file as the class method (`countries_list`)
needed to be defined before the tests and I wanted all the helper
functions to be together.

Whether methods are instance or class scoped is dependent on the test
need. The only method used outside of a test is `countries_list`, thus
this is the only method defined at a class level.

I've added some modest protection in the helpers so they will raise an
error if any of the lists accessed is empty. This has been done so we
don't have a chunk of tests defined based on a YAML list which is empty,
and thus never executes. In that scenario I'd expect the list and tests
to be removed.