Fix test response store for additional parameters
The flow test helper used a regular hash as the response store, however
the response store in FlowHelper is using a hash with indifferent
access for query parameters. This meant that the `get` method responds to
both string and symbols, however not in the tests. This broke the
ability to access additional parameter responses set in tests.