Remove unnecessary calculator methods
Removes unnecessary calculator methods from
MarriedCouplesAllowanceCalculator. These calculator are only used in
CalculateMarriedCouplesAllowanceFlow to determine the very next
question. They are not used in any of the views, any other calculations,
or even further down the flow as with `husband_income_measured?` which
is why that one has been retained.