Test core class of MaternityPaternityCalculatorFlow
This commit is the first in a series of commits to apply the ADR-3 [1]
style of flow tests to the MaternityPaternityCalculatorFlow.

This commit in particular adds tests to the one node that is defined in
MaternityPaternityCalculatorFlow. I've decided to break up the testing
of this flow into tests that match the individual files
(MaternityPaternityCalculatorFlow::AdoptionCalculatorFlow etc) that are
used to define it. I've done this as a way to try cope with the length
and complexity of the flows - as I found my initial pass, of putting
all the flow tests into a single file, created a file that was
overwhelming in length (> 2000 lines). The split isn't the easiest to
follow as all files are coupled and intertwined.

[1]: https://github.com/alphagov/smart-answers/blob/main/docs/arch/003-testing-flows.md