Helper for MaternityPaternityCalculatorFlow tests
This adds a helper file to reduce code duplication in the tests for
MaternityPaternityCalculatorFlow. It adds helpers to generate routes
through the common flows (maternity, paternity and both their adoption
equivalents).

The MaternityPaternityCalculatorFlow is a very complicated flow with
a large number of questions. Defining each answer for testing each node
becomes very repetitive and reduces the readability of the flow file. I
found the ambiguity of using a function to define the responses helped
to produce more readable flow tests.

This also acts as a way of defining in one place the complications in
going through the flows - where maternity adoption and paternity
adoption can be very different in types of questions; or how overseas
can make a big difference to the journey through maternity adoption.