Add tests for MaternityPaternityCalculatorFlow::AdoptionCalculatorFlow
This adds node based flow tests for the nodes defined in
MaternityPaternityCalculatorFlow::AdoptionCalculatorFlow.

The journey through this flow is complicated, there are a lot of nodes
and you can change from this flow to
SharedAdoptionMaternityPaterntiyFlow and back to this. You can also get
to this flow by starting a paternity flow and answering yes to an
adoption question.

One of the most confusing parts of this flow is the difference in
experience for when an adoption is from overseas, this causes a
difference in the ordering of questions that can be very hard to follow.
For example if you are adopting from overseas you can have the question
adoption_date_leave_starts? and 3 questions later have
adoption_is_the_employee_on_your_payroll?, whereas if you are not
adoption from overseas you answer
adoption_is_the_employee_on_your_payroll? 3 questions before
adoption_date_leave_starts?.

To try reduce the verbosity of the tests I've used a helper
`maternity_adoption_responses`. This allows the intent of the tests to
be easier to read, but comes at the expense of an extra step needed to
work out what responses were used. I decided to apply that function for
when there are more than 5 responses.