Add tests for MaternityPaternityCalculatorFlow::MaternityCalculatorFlow
This adds node based flow tests for the nodes defined in
MaternityPaternityCalculatorFlow::MaternityCalculatorFlow.

This flow follows a similar journey of questions to the Adoption and
Maternity equivalents. I assume this one was the first, as none of the
questions reference "maternity" in their name. Like the adoption tests
previously defined, I've used a helper `maternity_responses` to reduce
the verbosity of flow set-up.