Add tests for MaternityPaternityCalculatorFlow::PaternityCalculatorFlow
This adds node based flow tests for the nodes defined in
MaternityPaternityCalculatorFlow::PaternityCalculatorFlow.

This flow is bafflingly complicated and took a lot of time to try
understand how it is meant to behave. From question
`monthly_pay_paternity` onwards you can either be a paternity flow or an
adoption flow (including a maternity one) so the naming is rather
confused.