Add tests for SharedAdoptionMaternityPaternityFlow
This adds node based flow tests for the nodes defined in
MaternityPaternityCalculatorFlow::SharedAdoptionMaternityPaternityFlow.

Unlike the rest of MaternityPaternityCalculatorFlow, the flow here is
relatively easy to follow and thus quite easy to write readable tests
for. The journey to each of these nodes is quite consistent across the 3
types of flow (adoption, maternity, paternity) all joining this after
the equivalent earnings_for_pay_period question.

The use of helper methods for responses allowed these tests to be
concise.