Raise in Question::Date#range for unusable scenarios
Setting from and to for a date question now only has the affect of
configuring validation for the node. Therefore there should never be
scenarios where there is only one date and not the other or a from date
>= a to date, as these are not validatable.

This commit sets an exception to be raised if a range has been
misconfigured.