Rewrite flow tests for ChildBenefitTaxCalculatorFlow
This replaces the previous unit/integration flow tests with the node
based approach proposed in ADR 3 [1].

This flow is somewhat unusual in that it has a dynamic quantity of nodes
that are based on the answers provided to how_many_children? and
how_many_children_part_year? - where that could be up to 90 dynamic
nodes added. As the same code is iterated in every one of these
questions I've just put in a nominal test of one that exercises all the
logic.

[1]: https://github.com/alphagov/smart-answers/blob/main/docs/arch/003-testing-flows.md