Rewrite CheckUkVisaFlow tests
This rewrites the CheckUkVisaFlow test in the node style introduced in
ADR-3 [1].

CheckUkVisa is one of the most complicated Smart Answers (it ranks at
number 4 most complicated based on my crude filesize metric). There is a
lot of different routes through this flow and a daunting 38 different
outcome files.

The most complicated bit of logic, the purpose of visit, is repeated in
two questions - I therefore moved that into a helper file so it could be
re-used without engulfing the top of the file (as test helper methods
need to be static, they need to be defined before their usage).

The other challenging area is setting up the stubs for the world
locations so that individual scenarios will work. There seems to be a
quirk with the world locations stub where it can't be called multiple
times in a flow (presumably it stubs an index endpoint and subsequent
stubbings don't clear the previous one), so you can't adjust this
stubbing in particular test scenarios. Because of this I decided to have
one stub at the top of the file for all scenarios and set ivars for
country types. This carries with it an expense that all tests carry a
stubs they don't need (the stubs are surprisingly time expensive).

As a mark of the complexity of this I did find a number of unreachable
scenarios. I'll resolve those in a separate commit.

[1]: https://github.com/alphagov/smart-answers/blob/main/docs/arch/003-testing-flows.md