Remove unreachable transit outcome
It appears that every possible country you can enter will lead to the
question of `travelling_to_cta?` so the `outcome_no_visa_needed` node is
unreachable from this condition.

The question text too wouldn't make much sense without the
`travelling_to_cta?` question as it seems logical than anytime you
answer "Transit (on your way to somewhere else)" that it is followed
with a question of "where".