Rename full_time_tuition to tuition_maintenance_summary
This partial was only being called by outcome_uk_full_time_students.
However, it will soon be called by other outcomes, which may not
necessarily refer to 'full time'.

What this partial is _actually_ doing is presenting a summary of
the user's entitlement to tuition and maintenance loans. The only
part that is 'full time' is the URLs of the links in the partial,
which we will abstract away in a future commit.