Remove `uae_not_married` from `oru_result`
`uae_not_married` is not a valid country slug. Also it's not possible to
register a birth in UAE if the parents are not married, so even if this
conditional was amended to check a country slug of
`united-arab-emirates` and `paternity_declaration?` is true, this code
would still never be reached.