Remove check for DNA tests from oru_result
The only two countries that might ask for a DNA test are Libya and
Somalia. However it's not possible to register a birth in Libya or
Somalia as there is no embassy, so users are directed to
`no_embassy_result` instead rather than `oru_result`.

Therefore the guidance about DNA tests is never reached. The
accompanying DNA-related code isn't used anywhere else either so can be
removed.