Remove `north-korea` conditional from `oru_result`
This condition can never be reached. The case statement checks the
`registration_country`. `registration_country` checks the current or
country of birth in that order.

If the current country is North Korea, then the user is shown
`north_korea_result` not `oru_result`.

If the country of birth is North Korea and the user is in another
country, that other country will be set as the `registration_country`.
So it is impossible for `north-korea` to be the registration country" in
`oru_result`.