Remove logic from no_birth_certificate_result
This logic is unnecessary as the content for each country is the same
and this content can only be reached if the country of birth is in
`COUNTRIES_WITH_BIRTH_REGISTRATION_EXCEPTION` and the parents are not
married. That guard is handled in the flow, it doesn't need to be
repeated here.