DRY-up country list in Register a Birth flow tests
Rather than having a list of 50 countries, we can just get  the
countries we need from the lists in `RegistrationsDataQuery`.  We don't
need to use all of the country lists as some of them only relate to the
Register a Death flow.

Yemen has been replaced with Libya, as though neither have an embassy,
Yemen doesn't appear in any of the existing country lists.

Indonesia doesn't exist in any of the country lists either, but we need
to explicitly stub its location as it's required to test the logic in
one of the outcomes.