Remove the 'hide this page' JS component, stylesheet and test
The Find Support flow has been retired and the 'hide this page' button
and functionality is no longer required in Smart Answers.

We should remove the button, the button styling, functionality, 
configuration, tests and update the various views. But NOT any of the 
session-based flow code - which is still needed.

This commit removes the JavaScript component and test, the stylesheet,
and removes both JS and SCSS files from the Rails asset 'application'
files.