Remove hide_previous_answers_on_results_page
Only the find-coronavirus-support smart-answer used this functionality.
Now that that smart-answer has been retired we no longer need to keep
this check.

The setting hasn't been removed from find_coronavirus_support_flow as
that code is being removed in a separate PR and it would cause merge
conflicts to remove it here as well.