Change logic for UK Ancestry Visa's business information
There is an additional bullet point required for those coming to the UK
for business to be shown on the UK Ancestry Visa information. This was
previously set using a variable, but this will not be available when the
partial is loaded from the YAML configuration, so using the user's
response instead.