Remove before_pension_credit_date? condition
It's impossible not to satisfy the removed condition.
Pension credit age calculation for all genders is equivalent /
to the state pension date calculation for women. /
Gender stopped making a difference to the calculation in 1953. Women in 1953 reached state_pension_age in 2018.
This outcome partial is only rendered for users if they have not yet reached state pension date, so there is no need for the condition as the user would always be `before_pension_credit_date?`.