Convert flow tests for state_pension_through_partner
- converts the state pension through partner tests to use the node based testing
approach introduced in [this ADR](https://github.com/alphagov/smart-answers/blob/main/docs/arch/003-testing-flows.md)

I decided just to use 'your_pension_age_before_specific_date' and 'your_pension_age_after_specific_date' as opposed to 'on or before 5 April 2016' and 'on or after 6 April 2016' in the test descriptions. This was because I don't think the date itself is overly important to the tests, it made it easier to reuse code and made the test descriptions a bit easier to write. However I know this isn't representative to what the user will see so I am open to thoughts as to whether I should change this.