Use more than one payment option for country
For Austria, the payment options are different for `same_sex` vs
`opposite_sex` marriages. This commits enables different payment
partials to be used for different type of marriages. We could've just
left the payment info as hardcoded text in the `same_sex` partials but I
felt this was confusing as only one payment method is defined in the
marriage abroad YAML file for Austria [1]. Using this approach also DRYs
up the `same_sex` partials.

[1]: https://github.com/alphagov/smart-answers/blob/047f2eb4fdf0d45c6bfc05650c922fed2204e09e/config/smart_answers/marriage_abroad_services.yml#L275