Fix time sensitive test
This test started to fail when the date hit "01/10/2021" due to the
dynamic question `how_many_total_days?, which works out how many days
are available from 1st of October (start of leave year) to the current
date. As the response was 50 days, this test will fail for the next 50
days. Freezing time at "2021-12-01" made available enough days to comply
with the 50 day response.