Remove 2013-14 tax year in penalties calculator
We no longer need this tax year in the calculator and the tests have
been updated to test the 2014-15 tax year.

The test coverage has also been improved to ensure we are testing the
penalty charge bands. Moving `overdue_filing_days` and `overdue_payment_days`
out of the `private` scope allows us to check the number of overdue
days expected based on the dates provided. This should make it easier
to check the tests are correct in when we make similar updates in future.